Revert "added sword item and sword hit collision and more sounds"
This reverts commit c95cc1d73cd8c421d1c7aa6196cde16c52f1d4c5.

+++ Assets/Scripts/ArrowCollisionController.cs
@@ -2,7 +2,7 @@
 
 public class ArrowCollisionController : MonoBehaviour {
 
-    public float arrowDamage = 10.0f;
+    public int arrowDamage = 10;
 
     private void OnCollisionEnter(Collision collision)
     {
